Removed width 10000px style from events layout
Not sure why it was there. Fixed some problem in website-next that we don't have in website-honesly.
It was preventing the responsive layout from working in ie11. When changing screen size (below 960px breakpoint) the width 100% on the events div kicked in and then this went to 10000px because of the layout