Added the URL_BASENAME to the history manager
one on the client. If you need to test the code with a URL_BASENAME you can either wait for a build and try out the deployed version (the commit shah becomes the URL_BASENAME). Or you can test it in development by making the following changes to the Makefile, assuming the URL_BASENAME is ‘xxx’:
1. ln -fs ../../assets-honestly dist/static-site/xxx/assets-honestly
2. mkdir -p ./dist/static-site/xxx
Then run URL_BASENAME=xxx make dev-static