Events Section (#207)
* Workable Feed - Dynamic Content Spike

* Added join us page

* Changed the route for testing

* Import a snapshot of website-next

* Ported the website-next jobs feed into website-honestly

* disabling tests

* Lambda is not modern enough for this

* Fix compilation

* events page

* Parallel fetching for state

* Added events to site state

* Passed events state to component

* Added events to state to fix compile site test

* Used Promise props like getFeaturedPosts

* Changed css and js to absolute paths

* Added some styling

* Added more styles

* Removed tags from display

* Merge from master

* Changed event description to badgerBlack

* Moved hover underline to event title

* Moved the anchor inside h2 for border bottom

* Added badger brain env variable to example

* Overrode styles.css path in dev environment

The dev static site didn't like prepending / on styles. So created override in webpack dev and reverted webpack base config

* Used Link component instead of hard-coding

* Styled the events timeline title

* Added line-height

* Removed width 10000px style from events layout

Not sure why it was there. Fixed some problem in website-next that we don't have in website-honesly.
It was preventing the responsive layout from working in ie11. When changing screen size (below 960px breakpoint) the width 100% on the events div kicked in and then this went to 10000px because of the layout

* Added alt tag to event list images

* Added alt text to React Conf 2017 images

* Added the URL_BASENAME to the history manager

one on the client. If you need to test the code with a URL_BASENAME you can either wait for a build and try out the deployed version (the commit shah becomes the URL_BASENAME). Or you can test it in development by making the following changes to the Makefile, assuming the URL_BASENAME is ‘xxx’:
1. ln -fs ../../assets-honestly dist/static-site/xxx/assets-honestly
2. mkdir -p ./dist/static-site/xxx
Then run URL_BASENAME=xxx make dev-static

* Set events page background to badger white

* Adding scheduling for the publish function to run every night at 1am

* Adding comments to the cron job of the publish lambda