Enable setting specific links as WhatToReadNext links
This code change allows for setting specific links at the bottom
of a case study. Previously these were set randomly (ignoring the
current case study), which is still supported for when no specific
links are set or the number of specific links is less than 3.

The specific links will appear from left to right.