Add gif to baseServiceConfig asset rule
This makes the rule match the one already defined in the baseConfig
config. Previously this was interpreted as a new rule and would
cause all assets to be written to /dist/services/assets-honestly.
We already save all assets to /dist/assets-honestly via the baseConfig
rule so don't need to do it twice.

This fixes the issue that asset URLs were being written into the html
source code with wrong file hashes, causing them to 404 on the live
site (when not using JavaScript).