Fix flow error in timeline slice
Passing an `Array<{ image: string, year: number, Component:
ComponentType<{}> }>` into a component that expected `Array<{ image:
string, year: number }>` was throwing an error because flow thinks the
child component might modify the array to add an object that is missing
the `Component` field. Changing the child component’s props to expect a
`$ReadOnlyArray` instead fixed it as flow knows you’re not going to add
an item with missing fields to the array then.